requirements.txt
flask
mysql-connector-python
werkzeug
pyjwt
gunicorn
